refactor: 공통 -> 로컬 컴포넌트로 변경
